refactor: remove components file from CreateToken
