+++ src/components/CreateTokenForm/index.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const CreateTokenForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  align-items: center;
+`
+
+export default CreateTokenForm
