+++ src/pages/CreateToken/index.tsx
@@ -5,8 +5,8 @@ import { Button, Flex, useWalletModal } from '@koda-finance/summitswap-uikit'
 import styled from 'styled-components'
 import { Option } from 'react-dropdown'
 import { TranslateString } from 'utils/translateTextHelpers'
+import StyledDropdownWrapper from 'components/DropdownWrapper/StyledDropdownWrapper'
 import { TokenType, STANDARD_TOKEN_OPTION, LIQUIDITY_TOKEN_OPTION } from '../../constants/createToken'
-import { StyledDropdownWrapper } from './components'
 import { CreatedTokenDetails } from './types'
 import TokenDetails from './TokenDetails'
 import StandardTokenForm from './StandardTokenForm'
